requirements.txt
meilisearch
sqlalchemy
psycopg2-binary
python-dotenv
openai
pandas
matplotlib
fastapi 
uvicorn
matplotlib